refactor: create file for tab bar
