+++ app/src/main/java/com/example/todolistapp/Screen.kt
@@ -1,10 +1,7 @@
 package com.example.todolistapp
 
 sealed  class Screen(val route: String) {
-    object All : Screen("all")
-    object Completed : Screen("completed")
-}
-val bottomNavItems = listOf(
-    Screen.All,
-    Screen.Completed,
-)
+
+    data object All : Screen("all")
+    data object Completed : Screen("completed")
+}
